requirements.txt
streamlit
python-dotenv
plotly

groq

crewai
langchain
langchain-groq

chromadb
sentence-transformers
faiss-cpu

pandas
numpy
pypdf
docx2txt

yfinance
requests
beautifulsoup4
duckduckgo-search
tenacity
tiktoken
rich
pydantic

aiohttp
asyncio

loguru

pytest
pytest-asyncio